User-Perspective/Langchain/Reloaded/pyproject.toml
[project]
name = "reloaded"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "huggingface-hub>=0.33.0",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.15",
    "langchain-core>=0.3.65",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.0",
    "langchain-openai>=0.3.23",
    "numpy>=2.3.0",
    "openai>=1.86.0",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.7.0",
    "transformers>=4.52.4",
]
